CLOUD-80257 Restructure of user caching to make it work even for EnableGlobalMethodSecurity
