+++ core/src/main/java/com/sequenceiq/cloudbreak/cache/CachingConfig.java
@@ -27,12 +27,6 @@ import org.springframework.context.annotation.Configuration;
 @EnableAutoConfiguration
 public class CachingConfig implements CachingConfigurer {
 
-    public static final String TEMPORARY_AWS_CREDENTIAL_CACHE = "temporary_aws_credential";
-
-    private static final long TTL_IN_SECONDS = 5L * 60;
-
-    private static final long MAX_ENTRIES = 1000L;
-
     @Inject
     private List<CacheDefinition> cacheDefinitions;
