+++ core/src/main/java/com/sequenceiq/cloudbreak/conf/ScimAccountGroupReaderFilter.java
@@ -1,31 +1,42 @@
-package com.sequenceiq.cloudbreak.controller;
+package com.sequenceiq.cloudbreak.conf;
+
+import java.io.IOException;
 
 import javax.inject.Inject;
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.oauth2.provider.OAuth2Authentication;
 import org.springframework.stereotype.Service;
+import org.springframework.web.filter.OncePerRequestFilter;
 
 import com.sequenceiq.cloudbreak.domain.CbUser;
 import com.sequenceiq.cloudbreak.service.user.UserDetailsService;
 import com.sequenceiq.cloudbreak.service.user.UserFilterField;
 
 @Service
-public class AuthenticatedUserService {
+public class ScimAccountGroupReaderFilter extends OncePerRequestFilter {
 
     @Inject
     private UserDetailsService userDetailsService;
 
-    public CbUser getCbUser() {
+    // @SuppressWarnings("unchecked")
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException,
+            IOException {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         if (authentication != null) {
             OAuth2Authentication oauth = (OAuth2Authentication) authentication;
             if (oauth.getUserAuthentication() != null) {
                 String username = (String) authentication.getPrincipal();
-                return userDetailsService.getDetails(username, UserFilterField.USERNAME);
+                CbUser user = userDetailsService.getDetails(username, UserFilterField.USERNAME);
+                request.setAttribute("user", user);
             }
         }
-        return null;
+        filterChain.doFilter(request, response);
     }
 }
